Create Redirector_Setup (#298)
* Create Redirector_Setup

Guidelines on setting up Redirector to act as a YouTube-Invidious extension.

* Update Redirector_Setup

I have formally updated the list with a more comprehensive list of rules, including rules that address embeds and multiple domains at once. 

If anyone would like to try and simplify this list further, they are free to. Otherwise, I would consider this complete and ready for use.

* Update Redirector_Setup

* Rename Redirector_Setup to redirector_setup

* Rename redirector_setup to redirector_setup.md

* Enhancements and link to Redirector

* Small enhancement

Co-authored-by: TheFrenchGhosty <47571719+TheFrenchGhosty@users.noreply.github.com>