Rearranged some userscripts, removed outdated "Audio Mode" userscript, added additional details.
Userscripts that haven't been updated since the deprecation of the invidio.us domain have been moved to the bottom, with a note that they can be edited. "Invidious Audio Mode" userscript was removed as it's functionality is baked into Invidious now. Priority is given to userscripts updated as recently as 2023, with support for multiple different instances by default.